Fixed demo
