+++ cabbage_demo/__main__.py
@@ -2,37 +2,7 @@ import logging
 import sys
 
 import cabbage
-
-logger = logging.getLogger(__name__)
-task_manager = cabbage.TaskManager()
-
-
-@task_manager.task(queue="sums")
-def sum(a, b):
-    print(a + b)
-
-
-@task_manager.task(queue="sums")
-def sleep(i):
-    import time
-
-    time.sleep(i)
-
-
-@task_manager.task(queue="sums")
-def sum_plus_one(a, b):
-    print(a + b + 1)
-
-
-def client():
-    # sum.defer(a=3, b=5)
-    # sum.defer(a=5, b=7)
-    # sum.defer(a=5, b="a")
-    # sum_plus_one.defer(a=4, b=7)
-    # product.defer(a=2, b=8)
-    sleep.defer(lock="a", i=2)
-    sleep.defer(lock="a", i=3)
-    sleep.defer(lock="a", i=4)
+from cabbage_demo.cabbage_app import task_manager
 
 
 def main():
@@ -40,9 +10,15 @@ def main():
     process = sys.argv[1]
     if process == "worker":
         queue = sys.argv[2]
-        return cabbage.Worker(task_manager, queue).run()
+        worker = cabbage.Worker(
+            task_manager=task_manager, queue=queue, import_paths=["cabbage_demo.tasks"]
+        )
+        worker.run()
+
     else:
-        return client()
+        from cabbage_demo import client
+
+        return client.client()
 
 
 if __name__ == "__main__":
