+++ cabbage_demo/cabbage_app.py
@@ -1,0 +1,3 @@
+import cabbage
+
+task_manager = cabbage.TaskManager()
